# backend_rust/src/api_routes/portfolios.rs
use rocket::{State, serde::json::Json, http::Status};
use sqlx::SqlitePool;
use std::collections::HashMap;
use std::str::FromStr;
use rust_decimal::Decimal;
use crate::models::{Portfolio as DbPortfolio, Asset as DbAsset, Transaction as DbTransaction};
use crate::schemas::{AssetOut, PortfolioCreate, PortfolioOut, PortfolioUpdate, TransactionOut};

#[utoipa::path(
    post,
    path = "/api/portfolios/",
    responses(
        (status = 201, description = "Portfolio created", body = PortfolioOut),
        (status = 400, description = "Bad request")
    )
)]
#[post("/", data = "<portfolio>")]
pub async fn create_portfolio(
    pool: &State<SqlitePool>,
    portfolio: Json<PortfolioCreate>
) -> Result<Json<PortfolioOut>, Status> {
    let res = sqlx::query_as::<_, DbPortfolio>(
        "INSERT INTO portfolios (name, description, currency, base_currency) 
         VALUES (?, ?, ?, ?) RETURNING *"
    )
    .bind(&portfolio.name)
    .bind(&portfolio.description)
    .bind(&portfolio.currency)
    .bind(&portfolio.currency) // Using currency as base_currency default
    .fetch_one(pool.inner())
    .await;

    match res {
        Ok(p) => {
            let assets = fetch_assets_for_portfolio(pool.inner(), p.id).await
                .map_err(|_| Status::InternalServerError)?;
            Ok(Json(PortfolioOut {
                id: p.id,
                name: p.name,
                description: p.description,
                currency: p.currency,
                assets,
            }))
        },
        Err(_) => Err(Status::InternalServerError),
    }
}

async fn fetch_assets_for_portfolio(pool: &sqlx::sqlite::SqlitePool, portfolio_id: i32) -> Result<Vec<AssetOut>, sqlx::Error> {
    let assets = sqlx::query_as::<_, DbAsset>("SELECT * FROM assets WHERE portfolio_id = ?")
        .bind(portfolio_id)
        .fetch_all(pool)
        .await?;

    let asset_ids: Vec<i32> = assets.iter().map(|a| a.id).collect();

    let all_txs = if asset_ids.is_empty() {
        vec![]
    } else {
        let placeholders: Vec<String> = (0..asset_ids.len()).map(|_| "?".to_string()).collect();
        let sql = format!("SELECT * FROM transactions WHERE asset_id IN ({})", placeholders.join(","));
        let mut query = sqlx::query_as::<_, DbTransaction>(&sql);
        for id in &asset_ids {
            query = query.bind(id);
        }
        query.fetch_all(pool).await?
    };

    let mut txs_by_asset: HashMap<i32, Vec<TransactionOut>> = HashMap::new();
    for tx in all_txs {
        txs_by_asset
            .entry(tx.asset_id)
            .or_default()
            .push(TransactionOut {
                id: tx.id,
                asset_id: tx.asset_id,
                r#type: tx.r#type,
                quantity: Decimal::from_str(&tx.quantity).unwrap_or(Decimal::ZERO),
                price: Decimal::from_str(&tx.price).unwrap_or(Decimal::ZERO),
                fee: Decimal::from_str(&tx.fee).unwrap_or(Decimal::ZERO),
                date: tx.date,
            });
    }

    let out = assets.into_iter().map(|a| AssetOut {
        id: a.id,
        portfolio_id: a.portfolio_id,
        symbol: a.symbol,
        name: a.name,
        asset_type: a.asset_type,
        sector: a.sector,
        currency: a.currency,
        isin: a.isin,
        transactions: txs_by_asset.remove(&a.id).unwrap_or_default(),
    }).collect();

    Ok(out)
}

#[utoipa::path(
    get,
    path = "/api/portfolios/",
    responses(
        (status = 200, description = "List of portfolios", body = [PortfolioOut])
    )
)]
#[get("/")]
pub async fn list_portfolios(
pool: &State<SqlitePool>) -> Result<Json<Vec<PortfolioOut>>, Status> {
    let portfolios = sqlx::query_as::<_, DbPortfolio>("SELECT * FROM portfolios")
        .fetch_all(pool.inner())
        .await
        .map_err(|_| Status::InternalServerError)?;

    let all_assets = sqlx::query_as::<_, DbAsset>("SELECT * FROM assets")
        .fetch_all(pool.inner())
        .await
        .map_err(|_| Status::InternalServerError)?;

    let mut by_portfolio: HashMap<i32, Vec<AssetOut>> = all_assets
        .into_iter()
        .map(|a| AssetOut {
            id: a.id,
            portfolio_id: a.portfolio_id,
            symbol: a.symbol,
            name: a.name,
            asset_type: a.asset_type,
            sector: a.sector,
            currency: a.currency,
            isin: a.isin,
            transactions: vec![],
        })
        .fold(HashMap::new(), |mut map, asset| {
            map.entry(asset.portfolio_id).or_default().push(asset);
            map
        });

    let out = portfolios.into_iter().map(|p| PortfolioOut {
        id: p.id,
        name: p.name,
        description: p.description,
        currency: p.currency,
        assets: by_portfolio.remove(&p.id).unwrap_or_default(),
    }).collect();

    Ok(Json(out))
}

#[utoipa::path(
    get,
    path = "/api/portfolios/<id>",
    responses(
        (status = 200, description = "Portfolio found", body = PortfolioOut),
        (status = 404, description = "Portfolio not found")
    )
)]
#[get("/<id>")]
pub async fn get_portfolio(id: i32, pool: &State<SqlitePool>) -> Result<Json<PortfolioOut>, Status> {

    let p = sqlx::query_as::<_, DbPortfolio>("SELECT * FROM portfolios WHERE id = ?")
        .bind(id)
        .fetch_optional(pool.inner())
        .await
        .map_err(|_| Status::InternalServerError)?;

    match p {
        Some(p) => {
            let assets = fetch_assets_for_portfolio(pool.inner(), p.id).await
                .map_err(|_| Status::InternalServerError)?;
            Ok(Json(PortfolioOut {
                id: p.id,
                name: p.name,
                description: p.description,
                currency: p.currency,
                assets,
            }))
        },
        None => Err(Status::NotFound),
    }
}

#[utoipa::path(
    patch,
    path = "/api/portfolios/<id>",
    responses(
        (status = 200, description = "Portfolio updated", body = PortfolioOut),
        (status = 404, description = "Portfolio not found")
    )
)]
#[patch("/<id>", data = "<update>")]
pub async fn update_portfolio(
    id: i32,
    pool: &State<SqlitePool>,
    update: Json<PortfolioUpdate>
) -> Result<Json<PortfolioOut>, Status> {
    let new_currency = &update.currency;

    sqlx::query("UPDATE portfolios SET currency = ?, base_currency = ? WHERE id = ?")
        .bind(new_currency)
        .bind(new_currency)
        .bind(id)
        .execute(pool.inner())
        .await
        .map_err(|_| Status::InternalServerError)?;

    let p = sqlx::query_as::<_, DbPortfolio>("SELECT * FROM portfolios WHERE id = ?")
        .bind(id)
        .fetch_optional(pool.inner())
        .await
        .map_err(|_| Status::InternalServerError)?;

    match p {
        Some(p) => {
            let assets = fetch_assets_for_portfolio(pool.inner(), p.id).await
                .map_err(|_| Status::InternalServerError)?;
            Ok(Json(PortfolioOut {
                id: p.id,
                name: p.name,
                description: p.description,
                currency: p.currency,
                assets,
            }))
        },
        None => Err(Status::NotFound),
    }
}

#[utoipa::path(
    delete,
    path = "/api/portfolios/<id>",
    responses(
        (status = 204, description = "Portfolio deleted"),
        (status = 404, description = "Portfolio not found")
    )
)]
#[delete("/<id>")]
pub async fn delete_portfolio(id: i32, pool: &State<SqlitePool>) -> Result<Status, Status> {
    let portfolio = sqlx::query_as::<_, DbPortfolio>("SELECT * FROM portfolios WHERE id = ?")
        .bind(id)
        .fetch_optional(pool.inner())
        .await
        .map_err(|_| Status::InternalServerError)?;

    match portfolio {
        Some(_) => {},
        None => return Err(Status::NotFound),
    }

    let asset_count: (i64,) = sqlx::query_as("SELECT COUNT(*) FROM assets WHERE portfolio_id = ?")
        .bind(id)
        .fetch_one(pool.inner())
        .await
        .map_err(|_| Status::InternalServerError)?;

    if asset_count.0 > 0 {
        return Err(Status::Conflict);
    }

    let res = sqlx::query("DELETE FROM portfolios WHERE id = ?")
        .bind(id)
        .execute(pool.inner())
        .await
        .map_err(|_| Status::InternalServerError)?;

    if res.rows_affected() == 0 {
        return Err(Status::NotFound);
    }

    Ok(Status::NoContent)
}