# docs/openapi/openapi.json
{"components":{"schemas":{"Asset":{"properties":{"asset_type":{"type":"string"},"currency":{"type":"string"},"id":{"format":"int32","type":"integer"},"isin":{"nullable":true,"type":"string"},"name":{"type":"string"},"portfolio_id":{"format":"int32","type":"integer"},"sector":{"nullable":true,"type":"string"},"symbol":{"type":"string"}},"required":["id","portfolio_id","symbol","name","asset_type","currency"],"type":"object"},"AssetCreate":{"properties":{"asset_type":{"type":"string"},"currency":{"type":"string"},"isin":{"type":"string"},"name":{"type":"string"},"sector":{"nullable":true,"type":"string"},"symbol":{"type":"string"}},"required":["symbol","name","asset_type","currency"],"type":"object"},"AssetLookupResult":{"properties":{"asset_type":{"type":"string"},"name":{"type":"string"},"symbol":{"type":"string"}},"required":["symbol","name","asset_type"],"type":"object"},"AssetOut":{"properties":{"asset_type":{"type":"string"},"currency":{"type":"string"},"id":{"format":"int32","type":"integer"},"isin":{"nullable":true,"type":"string"},"name":{"type":"string"},"portfolio_id":{"format":"int32","type":"integer"},"sector":{"nullable":true,"type":"string"},"symbol":{"type":"string"},"transactions":{"items":{"$ref":"#/components/schemas/TransactionOut"},"type":"array"}},"required":["id","portfolio_id","symbol","name","asset_type","currency","transactions"],"type":"object"},"AssetTaxSummary":{"properties":{"asset_type":{"type":"string"},"average_cost":{"$ref":"#/components/schemas/Decimal"},"current_price":{"$ref":"#/components/schemas/Decimal"},"current_shares":{"$ref":"#/components/schemas/Decimal"},"market_value":{"$ref":"#/components/schemas/Decimal"},"realized_pnl":{"$ref":"#/components/schemas/Decimal"},"symbol":{"type":"string"},"tax_lots":{"items":{"$ref":"#/components/schemas/TaxLot"},"type":"array"},"total_cost":{"$ref":"#/components/schemas/Decimal"},"unrealized_pnl":{"$ref":"#/components/schemas/Decimal"},"unrealized_roi":{"$ref":"#/components/schemas/Decimal"}},"required":["symbol","asset_type","current_shares","average_cost","current_price","total_cost","market_value","unrealized_pnl","unrealized_roi","realized_pnl","tax_lots"],"type":"object"},"Portfolio":{"properties":{"base_currency":{"type":"string"},"currency":{"type":"string"},"description":{"nullable":true,"type":"string"},"id":{"format":"int32","type":"integer"},"name":{"type":"string"}},"required":["id","name","currency","base_currency"],"type":"object"},"PortfolioCreate":{"properties":{"currency":{"type":"string"},"description":{"nullable":true,"type":"string"},"name":{"type":"string"}},"required":["name","currency"],"type":"object"},"PortfolioOut":{"properties":{"assets":{"items":{"$ref":"#/components/schemas/AssetOut"},"type":"array"},"currency":{"type":"string"},"description":{"nullable":true,"type":"string"},"id":{"format":"int32","type":"integer"},"name":{"type":"string"}},"required":["id","name","currency","assets"],"type":"object"},"TaxLot":{"properties":{"buy_date":{"$ref":"#/components/schemas/DateTime"},"buy_price":{"$ref":"#/components/schemas/Decimal"},"latent_gain_loss":{"$ref":"#/components/schemas/Decimal"},"latent_roi":{"$ref":"#/components/schemas/Decimal"},"original_qty":{"$ref":"#/components/schemas/Decimal"},"remaining_qty":{"$ref":"#/components/schemas/Decimal"}},"required":["buy_date","buy_price","original_qty","remaining_qty","latent_gain_loss","latent_roi"],"type":"object"},"Transaction":{"properties":{"asset_id":{"format":"int32","type":"integer"},"date":{"$ref":"#/components/schemas/DateTime"},"fee":{"type":"string"},"id":{"format":"int32","type":"integer"},"price":{"type":"string"},"quantity":{"type":"string"},"type":{"type":"string"}},"required":["id","asset_id","type","quantity","price","fee","date"],"type":"object"},"TransactionCreate":{"properties":{"date":{"$ref":"#/components/schemas/DateTime"},"fee":{"$ref":"#/components/schemas/Decimal"},"price":{"$ref":"#/components/schemas/Decimal"},"quantity":{"$ref":"#/components/schemas/Decimal"},"type":{"type":"string"}},"required":["type","quantity","price","fee","date"],"type":"object"},"TransactionOut":{"properties":{"asset_id":{"format":"int32","type":"integer"},"date":{"$ref":"#/components/schemas/DateTime"},"fee":{"$ref":"#/components/schemas/Decimal"},"id":{"format":"int32","type":"integer"},"price":{"$ref":"#/components/schemas/Decimal"},"quantity":{"$ref":"#/components/schemas/Decimal"},"type":{"type":"string"}},"required":["id","asset_id","type","quantity","price","fee","date"],"type":"object"}}},"info":{"description":"","license":{"name":""},"title":"backend_rust","version":"0.1.0"},"openapi":"3.0.3","paths":{"/api/assets/<id>":{"delete":{"operationId":"delete_asset","parameters":[{"in":"path","name":"id","required":true,"schema":{"format":"int32","type":"integer"}}],"responses":{"204":{"description":"Asset deleted"},"404":{"description":"Asset not found"}},"tags":["crate::api_routes::transactions"]}},"/api/assets/lookup":{"get":{"operationId":"lookup_isin","parameters":[{"description":"ISIN to look up (12 alphanumeric characters)","in":"query","name":"isin","required":true,"schema":{"type":"string"}}],"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AssetLookupResult"}}},"description":"Asset metadata resolved from ISIN"},"400":{"description":"Invalid ISIN format"},"404":{"description":"ISIN not found"}},"tags":["crate::api_routes::lookup"]}},"/api/portfolios/":{"get":{"operationId":"list_portfolios","responses":{"200":{"content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/PortfolioOut"},"type":"array"}}},"description":"List of portfolios"}},"tags":["crate::api_routes::portfolios"]},"post":{"operationId":"create_portfolio","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PortfolioCreate"}}},"description":"","required":true},"responses":{"201":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PortfolioOut"}}},"description":"Portfolio created"},"400":{"description":"Bad request"}},"tags":["crate::api_routes::portfolios"]}},"/api/portfolios/<id>":{"delete":{"operationId":"delete_portfolio","parameters":[{"in":"path","name":"id","required":true,"schema":{"format":"int32","type":"integer"}}],"responses":{"204":{"description":"Portfolio deleted"},"404":{"description":"Portfolio not found"}},"tags":["crate::api_routes::portfolios"]},"get":{"operationId":"get_portfolio","parameters":[{"in":"path","name":"id","required":true,"schema":{"format":"int32","type":"integer"}}],"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PortfolioOut"}}},"description":"Portfolio found"},"404":{"description":"Portfolio not found"}},"tags":["crate::api_routes::portfolios"]},"patch":{"operationId":"update_portfolio","parameters":[{"in":"path","name":"id","required":true,"schema":{"format":"int32","type":"integer"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PortfolioUpdate"}}},"description":"","required":true},"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PortfolioOut"}}},"description":"Portfolio updated"},"404":{"description":"Portfolio not found"}},"tags":["crate::api_routes::portfolios"]}},"/api/portfolios/<id>/performance":{"get":{"operationId":"get_portfolio_performance","parameters":[{"in":"path","name":"id","required":true,"schema":{"format":"int32","type":"integer"}}],"responses":{"200":{"content":{"application/json":{"schema":{}}},"description":"Performance data retrieved"},"404":{"description":"Portfolio not found"}},"tags":["crate::api_routes::analytics"]}},"/api/portfolios/<id>/tax-summary":{"get":{"operationId":"get_portfolio_tax_summary","parameters":[{"in":"path","name":"id","required":true,"schema":{"format":"int32","type":"integer"}},{"in":"query","name":"strategy","required":false,"schema":{"nullable":true,"type":"string"}},{"in":"query","name":"threshold_days","required":false,"schema":{"format":"int64","nullable":true,"type":"integer"}}],"responses":{"200":{"content":{"application/json":{"schema":{}}},"description":"Tax summary retrieved"},"404":{"description":"Portfolio not found"}},"tags":["crate::api_routes::analytics"]}},"/api/portfolios/<portfolio_id>/assets":{"post":{"operationId":"create_asset","parameters":[{"in":"path","name":"portfolio_id","required":true,"schema":{"format":"int32","type":"integer"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AssetCreate"}}},"description":"","required":true},"responses":{"201":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AssetOut"}}},"description":"Asset created"},"400":{"description":"Asset already exists"},"404":{"description":"Portfolio not found"}},"tags":["crate::api_routes::transactions"]}},"/api/portfolios/<portfolio_id>/assets/<asset_id>/transactions":{"post":{"operationId":"create_transaction","parameters":[{"in":"path","name":"asset_id","required":true,"schema":{"format":"int32","type":"integer"}},{"in":"path","name":"portfolio_id","required":true,"schema":{"format":"int32","type":"integer"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TransactionCreate"}}},"description":"","required":true},"responses":{"201":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TransactionOut"}}},"description":"Transaction created"},"404":{"description":"Asset not found"}},"tags":["crate::api_routes::transactions"]}},"/api/portfolios/<portfolio_id>/transactions":{"get":{"operationId":"list_portfolio_transactions","parameters":[{"in":"path","name":"portfolio_id","required":true,"schema":{"format":"int32","type":"integer"}}],"responses":{"200":{"content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/TransactionOut"},"type":"array"}}},"description":"List of transactions"},"404":{"description":"Portfolio not found"}},"tags":["crate::api_routes::transactions"]}},"/api/transactions/<id>":{"delete":{"operationId":"delete_transaction","parameters":[{"in":"path","name":"id","required":true,"schema":{"format":"int32","type":"integer"}}],"responses":{"204":{"description":"Transaction deleted"},"404":{"description":"Transaction not found"}},"tags":["crate::api_routes::transactions"]}}}}